refactor(crawler): streamline code structure and enhance S3 bucket management
- Removed unused imports and functions to simplify the crawler logic.
- Integrated S3 bucket creation and management directly within the main function.
- Updated .gitignore to include additional download paths.
