+++ .gitignore
@@ -11,3 +11,4 @@ node_modules
 build
 package-lock.json
 downloads
+salim/downloads
